1318 - move cities folder to admin
