+++ src/components/admin/cities/grid/DeleteModal.tsx
